requirements.txt
absl-py>=0.9.0
astor>=0.8.1
cachetools>=4.1.0
certifi>=2020.4.5.1
chardet>=3.0.4
Click>=7.1.2
dlib>=19.19.0
face-recognition>=1.3.0
face-recognition-models>=0.3.0
gast>=0.3.3
google-auth>=1.16.0
google-auth-oauthlib>=0.4.1
google-pasta>=0.2.0
grpcio>=1.29.0
h5py>=2.10.0
idna>=2.9
Keras>=2.3.1
Keras-Applications>=1.0.8
Keras-Preprocessing>=1.1.2
Markdown>=3.2.2
numpy>=1.18.4
oauthlib>=3.1.0
opencv-python>=4.2.0.34
opt-einsum>=3.2.1
Pillow>=7.1.2
protobuf>=3.12.2
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
python-crontab>=2.5.1
python-dateutil>=2.8.1
PyYAML>=5.3.1
requests>=2.23.0
requests-oauthlib>=1.3.0
rsa>=4.0
scipy>=1.4.1
selenium>=3.141.0
six>=1.15.0
tensorboard>=2.2.2
tensorflow>=2.2.0
tensorflow-estimator>=2.2.0
termcolor>=1.1.0
urllib3>=1.25.9
Werkzeug>=1.0.1
wrapt>=1.12.1
